pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pictr"
version = "0.1.0"
description = "Pictr Storyboard Agent — Tubi MVP"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "httpx>=0.27,<1.0",   # used by TestClient in tests
    "python-dotenv>=1.0,<2.0",
    # Vertex AI SDK for Imagen image generation.
    # Requires: Vertex AI API enabled + billing on your GCP project.
    "google-cloud-aiplatform>=1.50,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.4,<1.0",
    "black>=24.4,<25.0",
    "mypy>=1.10,<2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

# ---------------------------------------------------------------------------
# Ruff — linter + import sorter
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line length handled by black formatter
]

# ---------------------------------------------------------------------------
# Black — formatter
# ---------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py311"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
